+++ CHANGELOG.md
@@ -8,6 +8,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Released]
 
+## [0.3.0] - 2021-02-11
+### Changed
+- Bump version to 0.3.0
+
+## [0.2.2] - 2021-02-10
+### Changed
+- Setup development env
+
+## [0.2.1] - 2021-02-09
+### Changed
+- Import styles
+### Added
+- Sass css file
+
 ## [0.2.0] - 2021-02-08
 ### Changed
 - Bump version to 0.2.0